+++ src/classes/babies/GappyBaby.ts
@@ -1,8 +1,8 @@
-import { CollectibleType } from "isaac-typescript-definitions";
+import { CollectibleType, SlotVariant } from "isaac-typescript-definitions";
 import {
   CallbackCustom,
-  isSlotMachine,
   ModCallbackCustom,
+  isSlotMachine,
 } from "isaacscript-common";
 import { mod } from "../../mod";
 import { Baby } from "../Baby";
@@ -12,6 +12,12 @@ export class GappyBaby extends Baby {
   @CallbackCustom(ModCallbackCustom.POST_SLOT_DESTROYED)
   postSlotDestroyed(slot: EntitySlot): void {
     if (!isSlotMachine(slot)) {
+      return;
+    }
+
+    // We blacklist Donation Machines since `isaacscript-common` thinks that they are destroyed when
+    // they are removed by Racing+.
+    if (slot.Variant === SlotVariant.DONATION_MACHINE) {
       return;
     }