+++ docs/history.md
@@ -5,6 +5,7 @@
 
 ### _v1.16.1_ - September 16th, 2023
 
+- Fixed the bug where having Soy Milk or Almond Milk would not cause "missed tear" style babies to be skipped. (Thanks sisuka)
 - Fixed bugs with the following babies:
   - Gem Baby (#237) - The effect now works properly. (Thanks sisuka)
   - Dark Space Soldier Baby (#398) - Will no longer be chosen if you have C Section. (Thanks sisuka)