+++ docs/history.md
@@ -7,10 +7,11 @@
 
 - Fixed the bug where having Soy Milk or Almond Milk would not cause "missed tear" style babies to be skipped. (Thanks sisuka)
 - Fixed bugs with the following babies:
+  - Water Baby (#3) - Will now have the proper +5 flat damage bonus to the tears. (Thanks sisuka)
   - Gappy Baby (#171) - Will no longer get free items in the shop. (Thanks sisuka)
   - Gem Baby (#237) - The effect now works properly. (Thanks sisuka)
   - Dark Space Soldier Baby (#398) - Will no longer be chosen if you have C Section. (Thanks sisuka)
-  - Tooth Head Baby (#442) - The effect no longer happens on every tear.
+  - Tooth Head Baby (#442) - The effect no longer happens on every tear. (Thanks sisuka)
 
 ### _v1.16.0_ - September 15th, 2023